backend/requirements.txt
fastapi
uvicorn[standard]
langgraph>=1,<2
langchain>=1,<2
langchain-google-genai
python-dotenv
sqlalchemy>=2,<3
psycopg[binary]>=3.1,<4
